chore(cleanup): moved code out of main.rs and combined tx ixs
